Likes box styles changed
